+++ src/main/java/org/folio/innreach/domain/service/ConfigurationService.java
@@ -1,11 +1,11 @@
 package org.folio.innreach.domain.service;
 
 import org.folio.innreach.domain.dto.folio.ResultList;
-import org.folio.innreach.domain.dto.folio.configuration.ConfigurationDTO;
+import org.folio.innreach.domain.dto.folio.circulation.CirculationSettingDTO;
 import org.springframework.stereotype.Service;
 
 @Service
 public interface ConfigurationService {
 
-    ResultList<ConfigurationDTO> fetchConfigurationsDetailsByModule(String module);
+    ResultList<CirculationSettingDTO> fetchCheckoutSettings();
 }