+++ src/main/java/org/folio/innreach/util/JsonHelper.java
@@ -2,11 +2,9 @@ package org.folio.innreach.util;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.google.gson.Gson;
-import com.google.gson.JsonObject;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
-import org.folio.innreach.domain.dto.folio.configuration.ConfigurationDTO;
+import org.folio.innreach.domain.dto.folio.circulation.CirculationSettingDTO;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
@@ -67,14 +65,16 @@ public class JsonHelper {
   }
 
 
-  public static Long getCheckoutTimeDurationInMilliseconds(List<ConfigurationDTO> configData) {
-    long checkOutTime = defaultCheckoutTimeDuration;
-    if(!configData.isEmpty()) {
-      var value = configData.get(0).getValue();
-      JsonObject valueObject = new Gson().fromJson(value, JsonObject.class);
-      checkOutTime = valueObject.get(CHECKOUT_TIMEOUT_DURATION).getAsLong();
-    }
-    return checkOutTime * 60000;
+  public static Long getCheckoutTimeDurationInMilliseconds(List<CirculationSettingDTO> configData) {
+    long duration = configData.stream()
+      .findFirst()
+      .map(CirculationSettingDTO::getValue)
+      .map(v -> v.get(CHECKOUT_TIMEOUT_DURATION))
+      .filter(Number.class::isInstance)
+      .map(Number.class::cast)
+      .map(Number::longValue)
+      .orElse(defaultCheckoutTimeDuration);
+    return duration * 60000;
   }
 
 }